requirements.txt
aniso8601==8.0.0
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
flask-cors==3.0.10
Flask-RESTful==0.3.8
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
pyasn1==0.4.8
python-dotenv==0.14.0
python-jose==3.2.0
pytz==2020.1
rethinkdb==2.4.7
rsa==4.6
six==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
